Update cache tests

+++ tests/testthat/helper-cache.R
@@ -1,24 +1,33 @@
-# Provides a test-safe cache location
-with_gutenberg_cache <- function(code) {
+# Provides test-safe cache locations
+with_gutenberg_cache <- function(code, type = "persistent") {
   # Mock the prompt inside the dlr package call to utils::askYesNo
   testthat::local_mocked_bindings(
     askYesNo = function(...) TRUE,
     .package = "utils"
   )
-  old <- dlr::app_cache_dir("gutenbergr")
 
-  tmp <- tempfile("gutenbergr-test-")
-  dir.create(tmp, recursive = TRUE)
+  old_type <- getOption("gutenbergr_cache_type")
+  old_path <- dlr::app_cache_dir("gutenbergr")
 
-  dlr::set_app_cache_dir("gutenbergr", cache_dir = tmp)
+  if (type == "persistent") {
+    tmp <- tempfile("gutenbergr-test-")
+    dir.create(tmp, recursive = TRUE)
+    options(gutenbergr_cache_type = "persistent")
+    dlr::set_app_cache_dir("gutenbergr", cache_dir = tmp)
 
-  on.exit(
-    {
-      dlr::set_app_cache_dir("gutenbergr", cache_dir = old)
+    on.exit({
+      options(gutenbergr_cache_type = old_type)
+      dlr::set_app_cache_dir("gutenbergr", cache_dir = old_path)
       unlink(tmp, recursive = TRUE)
-    },
-    add = TRUE
-  )
+    }, add = TRUE)
+  } else {
+    options(gutenbergr_cache_type = "session")
+
+    on.exit({
+      options(gutenbergr_cache_type = old_type)
+      dlr::set_app_cache_dir("gutenbergr", cache_dir = old_path)
+    }, add = TRUE)
+  }
 
   force(code)
 }
